Compact sizeof(HTMLAnchorElement) and sizeof(HTMLLinkElement) https://bugs.webkit.org/show_bug.cgi?id=206343
Reviewed by Ryosuke Niwa.

Data from Membuster is saying that HTMLAnchorElement is relatively frequently allocated
element (1252 elements are kept) while sizeof(HTMLAnchorElement) is not optimized well.
We also found that HTMLLinkElement is too large while it has so many paddings and opportunities
to optimize it. This patch shrinks them.

We also found that SharedStringHashHash's upper 32bit is always zero, and nobody is using this
effectively. We make it 32bit.

No behavior change.

* html/HTMLAnchorElement.h:
* html/HTMLLinkElement.cpp:
(WebCore::HTMLLinkElement::HTMLLinkElement):
(WebCore::HTMLLinkElement::process):
* html/HTMLLinkElement.h:
* html/LinkIconCollector.h:
* html/LinkIconType.h:
* html/LinkRelAttribute.cpp:
(WebCore::LinkRelAttribute::LinkRelAttribute):
* html/LinkRelAttribute.h:
(): Deleted.
* page/Page.h:
* page/VisitedLinkStore.h:
* platform/SharedStringHash.h:
(WebCore::SharedStringHashHash::avoidDeletedValue): Deleted.

Canonical link: https://commits.webkit.org/219485@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254739 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/page/Page.h
@@ -147,7 +147,7 @@ class VisitedLinkStore;
 class WebGLStateTracker;
 class WheelEventDeltaFilter;
 
-using SharedStringHash = uint64_t;
+using SharedStringHash = uint32_t;
 
 enum class CanWrap : bool;
 enum class DidWrap : bool;
